Significant linting and chopping the views file into more manageable chunks for further work
