+++ qcrbox_frontend/qcrbox/urls.py
@@ -1,11 +1,11 @@
 '''QCRBox URL Configuration
 
-The `urlpatterns` list routes URLs to views. For more information please see:
+The `urlpatterns` list routes URLs to  For more information please see:
     https://docs.djangoproject.com/en/4.1/topics/http/urls/
 Examples:
 Function views
     1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
+    2. Add a URL to urlpatterns:  path('', home, name='home')
 Class-based views
     1. Add an import:  from other_app.views import Home
     2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
@@ -18,39 +18,39 @@ from django.urls import path
 from django.conf import settings
 from django.conf.urls.static import static
 
-from . import views
+from qcrbox.views import datasets, debug, groups, users, workflows
 
 urlpatterns = [
-    path('', views.landing, name='landing'),
-    path('workflow', views.initialise_workflow, name='initialise_workflow'),
-    path('workflow/<file_id>', views.workflow, name='workflow'),
+    path('', workflows.landing, name='landing'),
+    path('workflow', workflows.initialise_workflow, name='initialise_workflow'),
+    path('workflow/<file_id>', workflows.workflow, name='workflow'),
     path(
         'workflow/<file_id>/pending/<command_id>',
-        views.workflow_pending,
+        workflows.workflow_pending,
         name='workflow-pending',
     ),
-    path('download/<file_id>', views.download, name='download'),
+    path('download/<file_id>', datasets.download, name='download'),
 
-    path('login', views.login_view, name='login'),
-    path('logout', views.logout_view, name='logout'),
+    path('login', users.login_view, name='login'),
+    path('logout', users.logout_view, name='logout'),
 
-    path('view_users', views.view_users, name='view_users'),
-    path('create_user', views.create_user, name='create_user'),
-    path('edit_user/<user_id>', views.update_user, name='edit_user'),
-    path('edit_account/', views.edit_user, name='edit_account'),
-    path('edit_password/', views.update_password, name='edit_password'),
-    path('delete_user/<user_id>', views.delete_user, name='delete_user'),
+    path('view_users', users.view_users, name='view_users'),
+    path('create_user', users.create_user, name='create_user'),
+    path('edit_user/<user_id>', users.update_user, name='edit_user'),
+    path('edit_account/', users.edit_user, name='edit_account'),
+    path('edit_password/', users.update_password, name='edit_password'),
+    path('delete_user/<user_id>', users.delete_user, name='delete_user'),
 
-    path('create_group', views.create_group, name='create_group'),
-    path('view_groups', views.view_groups, name='view_groups'),
-    path('edit_group/<group_id>', views.update_group, name='edit_group'),
-    path('delete_group/<group_id>', views.delete_group, name='delete_group'),
+    path('create_group', groups.create_group, name='create_group'),
+    path('view_groups', groups.view_groups, name='view_groups'),
+    path('edit_group/<group_id>', groups.update_group, name='edit_group'),
+    path('delete_group/<group_id>', groups.delete_group, name='delete_group'),
 
-    path('data_history/<dataset_id>', views.history_dashboard, name='dataset_history'),
-    path('view_datasets', views.view_datasets, name='view_datasets'),
-    path('delete_dataset/<dataset_id>', views.delete_dataset, name='delete_dataset'),
-    path('visualise/<dataset_id>', views.visualise, name='visualise'),
+    path('data_history/<dataset_id>', datasets.history_dashboard, name='dataset_history'),
+    path('view_datasets', datasets.view_datasets, name='view_datasets'),
+    path('delete_dataset/<dataset_id>', datasets.delete_dataset, name='delete_dataset'),
+    path('visualise/<dataset_id>', datasets.visualise, name='visualise'),
 
-    path('frontend_logs', views.frontend_logs, name='frontend_logs'),
+    path('frontend_logs', debug.frontend_logs, name='frontend_logs'),
 
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
