+++ qcrbox_frontend/qcrbox/views/generic.py
